tests: Refactor e2e tests
