+++ test/e2e/features/ec2/types.go
@@ -5,7 +5,6 @@ package ec2
 import (
 	"path/filepath"
 
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/fsutils"
@@ -17,8 +16,6 @@ const (
 	routeName      = "ec2-route"
 	backendName    = "ec2-backend"
 	awsCliPodName  = "aws-cli-ec2"
-	localstackNS   = "localstack"
-	localstackSvc  = "localstack"
 	ec2Region      = "us-east-1"
 	ec2Port        = 8080
 	ec2TagApp      = "payments"
@@ -30,13 +27,6 @@ var (
 	setupManifest      = filepath.Join(fsutils.MustGetThisDir(), "testdata", "setup.yaml")
 	awsCliPodManifest  = filepath.Join(fsutils.MustGetThisDir(), "testdata", "aws-cli.yaml")
 	ec2BackendManifest = filepath.Join(fsutils.MustGetThisDir(), "testdata", "ec2-backend.yaml")
-
-	localstackService = corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      localstackSvc,
-			Namespace: localstackNS,
-		},
-	}
 
 	proxyObjectMeta = metav1.ObjectMeta{
 		Name:      gatewayName,
