+++ test/e2e/tests/automtls_istio_test.go
@@ -1,89 +1,15 @@
 //go:build e2e
 
-package tests_test
+package tests
 
 import (
-	"context"
-	"os"
 	"testing"
 
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e"
-	. "github.com/kgateway-dev/kgateway/v2/test/e2e/tests"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/install"
-	testruntime "github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/runtime"
-	"github.com/kgateway-dev/kgateway/v2/test/testutils"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e/tests/automtlsistio"
 )
 
-const (
-	minAutomtlsIstioVersion = "1.24.2"
-)
-
-// TestKgatewayIstioAutoMtls is the function which executes a series of tests against a given installation
+// TestKgatewayIstioAutoMtls tests Istio auto-mTLS integration.
 func TestKgatewayIstioAutoMtls(t *testing.T) {
-	ctx := context.Background()
-
-	// Set Istio version if not already set
-	if os.Getenv(testruntime.IstioVersionEnv) == "" {
-		os.Setenv(testruntime.IstioVersionEnv, testruntime.DefaultIstioVersion) // Using default istio version
-	}
-
-	if testruntime.ShouldSkipIstioVersion(t, minAutomtlsIstioVersion) {
-		t.Skip("Skipping due to https://github.com/istio/istio/issues/53846 which is fixed in Istio >= 1.24.2")
-		return
-	}
-
-	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "automtls-istio-test")
-	testInstallation := e2e.CreateTestInstallation(
-		t,
-		&install.Context{
-			InstallNamespace:          installNs,
-			ProfileValuesManifestFile: e2e.CommonRecommendationManifest,
-			ValuesManifestFile:        e2e.ManifestPath("istio-automtls-enabled-helm.yaml"),
-		},
-	)
-
-	err := testInstallation.AddIstioctl(ctx)
-	if err != nil {
-		t.Errorf("failed to add istioctl: %v\n", err)
-	}
-
-	// Set the env to the install namespace if it is not already set
-	if !nsEnvPredefined {
-		os.Setenv(testutils.InstallNamespace, installNs)
-	}
-
-	// We register the cleanup function _before_ we actually perform the installation.
-	// This allows us to uninstall kgateway, in case the original installation only completed partially
-	testutils.Cleanup(t, func() {
-		if !nsEnvPredefined {
-			os.Unsetenv(testutils.InstallNamespace)
-		}
-		if t.Failed() {
-			testInstallation.PreFailHandler(ctx, t)
-
-			// Generate istioctl bug report
-			testInstallation.CreateIstioBugReport(ctx)
-		}
-
-		// Uninstall kgateway
-		testInstallation.UninstallKgateway(ctx, t)
-
-		// Uninstall Istio
-		err = testInstallation.UninstallIstio()
-		if err != nil {
-			t.Errorf("failed to add istioctl: %v\n", err)
-		}
-	})
-
-	// Install Istio before kgateway to make sure istiod is present before kgateway for sds
-	err = testInstallation.InstallMinimalIstio(ctx)
-	if err != nil {
-		t.Errorf("failed to add istioctl: %v\n", err)
-	}
-
-	// Install kgateway
-	testInstallation.InstallKgatewayFromLocalChart(ctx, t)
-
-	AutomtlsIstioSuiteRunner().Run(ctx, t, testInstallation)
+	automtlsistio.Run(t, e2e.DefaultInstallationFactory)
 }
