+++ test/e2e/tests/routereplacement/routereplacement.go
@@ -1,0 +1,66 @@
+//go:build e2e
+
+// Package routereplacement holds the body of the TestRouteReplacement e2e
+// test.
+package routereplacement
+
+import (
+	"context"
+	"os"
+	"testing"
+
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e"
+	routereplacementfeature "github.com/kgateway-dev/kgateway/v2/test/e2e/features/routereplacement"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/install"
+	"github.com/kgateway-dev/kgateway/v2/test/testutils"
+)
+
+// Run executes the TestRouteReplacement scenario against the Installation
+// produced by the supplied factory.
+func Run(t *testing.T, factory e2e.InstallationFactory) {
+	ctx := t.Context()
+	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "route-replacement-test")
+	testInstallation := factory(
+		t,
+		&install.Context{
+			InstallNamespace:          installNs,
+			ProfileValuesManifestFile: e2e.CommonRecommendationManifest,
+			ValuesManifestFile:        e2e.EmptyValuesManifestPath,
+			ExtraHelmArgs: []string{
+				"--set", "validation.level=strict",
+			},
+		},
+	)
+
+	// Set the env to the install namespace if it is not already set
+	if !nsEnvPredefined {
+		os.Setenv(testutils.InstallNamespace, installNs)
+	}
+
+	// We register the cleanup function _before_ we actually perform the installation.
+	// This allows us to uninstall kgateway, in case the original installation only completed partially
+	testutils.Cleanup(t, func() {
+		ctx := context.Background() // t.Context() is canceled before t's cleanup runs
+		if !nsEnvPredefined {
+			os.Unsetenv(testutils.InstallNamespace)
+		}
+		if t.Failed() {
+			testInstallation.PreFailHandler(ctx, t)
+		}
+
+		testInstallation.UninstallKgateway(ctx, t)
+	})
+
+	// Install kgateway
+	testInstallation.InstallKgatewayFromLocalChart(ctx, t)
+
+	SuiteRunner().Run(ctx, t, testInstallation.Underlying())
+}
+
+// SuiteRunner returns the suite runner for the TestRouteReplacement scenario.
+func SuiteRunner() e2e.SuiteRunner {
+	routeReplacementSuiteRunner := e2e.NewSuiteRunner(false)
+	routeReplacementSuiteRunner.Register("RouteReplacement", routereplacementfeature.NewTestingSuite)
+	return routeReplacementSuiteRunner
+}
