+++ test/e2e/tests/validation_test.go
@@ -1,54 +1,15 @@
 //go:build e2e
 
-package tests_test
+package tests
 
 import (
-	"context"
-	"os"
 	"testing"
 
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e"
-	. "github.com/kgateway-dev/kgateway/v2/test/e2e/tests"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/install"
-	"github.com/kgateway-dev/kgateway/v2/test/testutils"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e/tests/routereplacement"
 )
 
+// TestRouteReplacement tests route replacement under strict validation.
 func TestRouteReplacement(t *testing.T) {
-	ctx := context.Background()
-	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "route-replacement-test")
-	testInstallation := e2e.CreateTestInstallation(
-		t,
-		&install.Context{
-			InstallNamespace:          installNs,
-			ProfileValuesManifestFile: e2e.CommonRecommendationManifest,
-			ValuesManifestFile:        e2e.EmptyValuesManifestPath,
-			ExtraHelmArgs: []string{
-				"--set", "validation.level=strict",
-			},
-		},
-	)
-
-	// Set the env to the install namespace if it is not already set
-	if !nsEnvPredefined {
-		os.Setenv(testutils.InstallNamespace, installNs)
-	}
-
-	// We register the cleanup function _before_ we actually perform the installation.
-	// This allows us to uninstall kgateway, in case the original installation only completed partially
-	testutils.Cleanup(t, func() {
-		if !nsEnvPredefined {
-			os.Unsetenv(testutils.InstallNamespace)
-		}
-		if t.Failed() {
-			testInstallation.PreFailHandler(ctx, t)
-		}
-
-		testInstallation.UninstallKgateway(ctx, t)
-	})
-
-	// Install kgateway
-	testInstallation.InstallKgatewayFromLocalChart(ctx, t)
-
-	RouteReplacementSuiteRunner().Run(ctx, t, testInstallation)
+	routereplacement.Run(t, e2e.DefaultInstallationFactory)
 }
