+++ test/e2e/tests/zero_downtime_test.go
@@ -1,63 +1,15 @@
 //go:build e2e
 
-package tests_test
+package tests
 
 import (
-	"context"
-	"os"
-	"path/filepath"
 	"testing"
 
-	"k8s.io/apimachinery/pkg/types"
-
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
-	"github.com/kgateway-dev/kgateway/v2/pkg/utils/fsutils"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/common"
-	. "github.com/kgateway-dev/kgateway/v2/test/e2e/tests"
-	"github.com/kgateway-dev/kgateway/v2/test/e2e/testutils/install"
-	"github.com/kgateway-dev/kgateway/v2/test/testutils"
+	"github.com/kgateway-dev/kgateway/v2/test/e2e/tests/zerodowntime"
 )
 
+// TestZeroDowntimeRollout tests that proxy rollouts don't drop traffic.
 func TestZeroDowntimeRollout(t *testing.T) {
-	ctx := context.Background()
-	installNs, nsEnvPredefined := envutils.LookupOrDefault(testutils.InstallNamespace, "zero-downtime")
-	testInstallation := e2e.CreateTestInstallation(
-		t,
-		&install.Context{
-			InstallNamespace:          installNs,
-			ProfileValuesManifestFile: e2e.CommonRecommendationManifest,
-			ValuesManifestFile:        e2e.EmptyValuesManifestPath,
-		},
-	)
-
-	// Set the env to the install namespace if it is not already set.
-	if !nsEnvPredefined {
-		os.Setenv(testutils.InstallNamespace, installNs)
-	}
-
-	// We register the cleanup function _before_ we actually perform the installation.
-	// This allows us to uninstall, in case the original installation only completed partially.
-	testutils.Cleanup(t, func() {
-		if !nsEnvPredefined {
-			os.Unsetenv(testutils.InstallNamespace)
-		}
-		if t.Failed() {
-			testInstallation.PreFailHandler(ctx, t)
-		}
-
-		testInstallation.UninstallKgateway(ctx, t)
-	})
-
-	testInstallation.InstallKgatewayFromLocalChart(ctx, t)
-
-	common.SetupBaseConfig(ctx, t, testInstallation,
-		filepath.Join(fsutils.MustGetThisDir(), "../features/zero_downtime_rollout/testdata", "gateway.yaml"),
-	)
-	common.SetupBaseGateway(ctx, t, testInstallation, types.NamespacedName{
-		Namespace: "default",
-		Name:      "gw",
-	})
-
-	ZeroDowntimeRolloutSuiteRunner().Run(ctx, t, testInstallation)
+	zerodowntime.Run(t, e2e.DefaultInstallationFactory)
 }
